refactor(p2): extract ChatListSubViews into sibling file (stage 3 of 4)
Moves the 3 sub-composables (parent L1360-1642 of post-stage-2
ChatList.kt) into the new sibling ChatListSubViews.kt in the same
package: ChatListPreview (search preview mode), ChatSuggestionsRow
(suggestion chip strip), BoxScope.MessageJumper (floating jump-to
navigator). Together they implement the timeline's accessory views.

Byte-equal: bodies and signatures unchanged. Inline FQNs in
ChatListPreview body (me.rerere.ai.core.MessageRole.USER,
androidx.compose.foundation.shape.RoundedCornerShape,
androidx.compose.foundation.BorderStroke) preserved verbatim.

Visibility widenings (3 sibling p→i, for the now-cross-sibling callers
still in coord — ChatList entry calls ChatListPreview, ChatListNormal
calls ChatSuggestionsRow + BoxScope.MessageJumper):
ChatListPreview, ChatSuggestionsRow, BoxScope.MessageJumper.

No coord visibility widenings (Stage 3 sub-composables only reach back
to Stage 1's Support sibling for extractMatchingSnippet + buildHighlightedText,
both already widened to internal in Stage 1).

Coordinator strip: 1642 -> 1342 (-300 with import cleanup).
Imports trimmed: 16 orphans — Lucide + 4 arrow/chevron icons
(ArrowDownToLine, ArrowUpToLine, ChevronDown, ChevronUp), 2 slide
animations (slideInHorizontally, slideOutHorizontally), combinedClickable,
IntrinsicSize, LazyRow, HorizontalDivider, surfaceColorAtElevation,
clip, fastCoerceAtLeast, WorkspaceSearchField, utils.plus — all
verified body refs only in extracted range.

Pre-existing dead imports preserved verbatim.

Baseline 3/467 unit tests preserved.
